refactor: Adjust namespaces and add ctor to Timeouts
