+++ src/Dodo.HttpClient.ResiliencePolicies.Tests/HttpClientBuilderExtensionsTests.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Net.Http;
 using Dodo.HttpClientResiliencePolicies.Tests.Fakes;
+using Dodo.HttpClientResiliencePolicies.TimeoutPolicy;
 using Microsoft.Extensions.DependencyInjection;
 using NUnit.Framework;
 
@@ -65,7 +66,7 @@ namespace Dodo.HttpClientResiliencePolicies.Tests
 			serviceCollection.AddJsonClient<IMockJsonClient, MockJsonClient>(
 				new Uri("http://example.com/"),
 				c => {
-					c.OverallTimeoutPolicySettings.Timeout = overallTimeout;
+					c.OverallTimeoutPolicySettings = new OverallTimeoutPolicySettings(overallTimeout);
 				});
 
 			var services = serviceCollection.BuildServiceProvider();
