+++ src/Dodo.HttpClient.ResiliencePolicies/ResiliencePoliciesSettings.cs
@@ -1,6 +1,6 @@
-using Dodo.HttpClientResiliencePolicies.CircuitBreakerSettings;
-using Dodo.HttpClientResiliencePolicies.RetrySettings;
-using Dodo.HttpClientResiliencePolicies.TimeoutPolicySettings;
+using Dodo.HttpClientResiliencePolicies.CircuitBreakerPolicy;
+using Dodo.HttpClientResiliencePolicies.RetryPolicy;
+using Dodo.HttpClientResiliencePolicies.TimeoutPolicy;
 
 namespace Dodo.HttpClientResiliencePolicies
 {
