+++ src/Dodo.HttpClient.ResiliencePolicies/TimeoutPolicy/ITimeoutPolicySettings.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace Dodo.HttpClientResiliencePolicies.TimeoutPolicy
+{
+	public interface ITimeoutPolicySettings
+	{
+		TimeSpan Timeout { get; }
+	}
+}
