--- src/Dodo.HttpClient.ResiliencePolicies/TimeoutPolicySettings/ITimeoutPolicySettings.cs
@@ -1,11 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Dodo.HttpClientResiliencePolicies.TimeoutPolicySettings
-{
-	public interface ITimeoutPolicySettings
-	{
-		TimeSpan Timeout { get; set; }
-	}
-}
